+++ items/itemdef.cpp
@@ -6,6 +6,6 @@ ItemDef::ItemDef(const ConfigNode& config){
 
 }
 
-int ItemDef::use(Player& owner, ItemStack& itemStack, const Screen& screen){
+double ItemDef::use(Player& owner, ItemStack& itemStack, const Screen& screen){
 	return 0;
 }
