+++ items/itemdef.h
@@ -15,6 +15,6 @@ class ItemDef{
 public:
 	int ID;
 	ItemDef(const ConfigNode& config);
-	virtual int use(Player& owner, ItemStack& item, const Screen& screen);
+	virtual double use(Player& owner, ItemStack& item, const Screen& screen);
 	virtual Block* asBlock(){return nullptr;}
 };
