+++ items/tool.h
@@ -9,7 +9,7 @@ class Tool : public ItemDef{
 public:
 	Tool(const ConfigNode& config);
 
-	virtual int use(Player& owner, ItemStack& item, const Screen& screen);
+	virtual double use(Player& owner, ItemStack& item, const Screen& screen);
 private:
 	Block* mineAt(World* world, Vector2i pos, int layer);
 };
